refactor ui code

+++ src/graphical/settings.rs
@@ -1,8 +1,8 @@
 // the main struct, that stores all settings data needed for drawing the UI
 pub struct Settings {
-    pub shutter_angle: f32,
-    pub iso: f32,
-    pub fps: f32,
+    pub shutter_angle: f64,
+    pub iso: f64,
+    pub fps: f64,
     pub recording_format: RecordingFormat,
     pub grid: Grid,
 }
